Move mouse wheel scrolling to separate file
