+++ compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/gestures/Scrollable.kt
@@ -16,8 +16,9 @@
 
 package androidx.compose.foundation.gestures
 
-import androidx.compose.foundation.MutatePriority as FoundationMutatePriority
-import androidx.compose.animation.core.*
+import androidx.compose.animation.core.AnimationState
+import androidx.compose.animation.core.DecayAnimationSpec
+import androidx.compose.animation.core.animateDecay
 import androidx.compose.animation.rememberSplineBasedDecay
 import androidx.compose.foundation.*
 import androidx.compose.foundation.gestures.Orientation.Horizontal
@@ -36,20 +37,12 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.input.pointer.*
 import androidx.compose.ui.modifier.ModifierLocalProvider
 import androidx.compose.ui.modifier.modifierLocalOf
-import androidx.compose.ui.node.DelegatingNode
-import androidx.compose.ui.node.ModifierNodeElement
-import androidx.compose.ui.node.PointerInputModifierNode
-import androidx.compose.ui.platform.InspectorInfo
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.platform.debugInspectorInfo
 import androidx.compose.ui.unit.*
-import androidx.compose.ui.util.fastAny
-import androidx.compose.ui.util.fastForEach
 import kotlin.math.abs
-import kotlin.math.roundToInt
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
@@ -265,6 +258,7 @@ private fun Modifier.pointerScrollable(
     }
     val draggableState = remember { ScrollDraggableState(scrollLogic) }
     val scrollConfig = platformScrollConfig()
+    val density = LocalDensity.current.density
 
     return this
         .then(DraggableElement(
@@ -282,245 +276,15 @@ private fun Modifier.pointerScrollable(
             },
             canDrag = { down -> down.type != PointerType.Mouse }
         ))
-        .then(MouseWheelScrollElement(scrollLogic, scrollConfig))
+        .then(MouseWheelScrollableElement(scrollLogic, scrollConfig, density))
         .nestedScroll(nestedScrollConnection, nestedScrollDispatcher.value)
 }
 
 // {} isn't being memoized for us, so extract this to make sure we compare equally on recomposition.
 private val NoOpOnDragStarted: suspend CoroutineScope.(startedPosition: Offset) -> Unit = {}
 
-private class MouseWheelScrollElement(
-    val scrollingLogicState: State<ScrollingLogic>,
-    val mouseWheelScrollConfig: ScrollConfig
-) : ModifierNodeElement<MouseWheelScrollNode>() {
-    override fun create(): MouseWheelScrollNode {
-        return if (mouseWheelScrollConfig.isSmoothScrollingEnabled) {
-            AnimatedMouseWheelScrollNode(scrollingLogicState, mouseWheelScrollConfig)
-        } else {
-            RawMouseWheelScrollNode(scrollingLogicState, mouseWheelScrollConfig)
-        }
-    }
-
-    override fun update(node: MouseWheelScrollNode) {
-        node.scrollingLogicState = scrollingLogicState
-        node.mouseWheelScrollConfig = mouseWheelScrollConfig
-    }
-
-    override fun hashCode(): Int {
-        var result = scrollingLogicState.hashCode()
-        result = 31 * result + mouseWheelScrollConfig.hashCode()
-        return result
-    }
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is MouseWheelScrollElement) return false
-
-        if (scrollingLogicState != other.scrollingLogicState) return false
-        if (mouseWheelScrollConfig != other.mouseWheelScrollConfig) return false
-        return true
-    }
-
-    override fun InspectorInfo.inspectableProperties() = Unit
-}
-
-// TODO(levima) Save the ScrollingLogic value in the ScrollableNode so we won't need a State here.
-private abstract class MouseWheelScrollNode(
-    var scrollingLogicState: State<ScrollingLogic>,
-    var mouseWheelScrollConfig: ScrollConfig
-) : DelegatingNode(), PointerInputModifierNode {
-    protected val scrollingLogic get() = scrollingLogicState.value
-
-    private val pointerInputNode = delegate(SuspendingPointerInputModifierNode {
-        mouseWheelInput()
-    })
-
-    override fun onPointerEvent(
-        pointerEvent: PointerEvent,
-        pass: PointerEventPass,
-        bounds: IntSize
-    ) {
-        pointerInputNode.onPointerEvent(pointerEvent, pass, bounds)
-    }
-
-    override fun onCancelPointerInput() {
-        pointerInputNode.onCancelPointerInput()
-    }
-
-    private suspend fun PointerInputScope.mouseWheelInput() = awaitPointerEventScope {
-        while (true) {
-            val event = awaitScrollEvent()
-            if (!event.isConsumed) {
-                val consumed = onMouseWheel(event)
-                if (consumed) {
-                    event.consume()
-                }
-            }
-        }
-    }
-
-    protected abstract fun PointerInputScope.onMouseWheel(pointerEvent: PointerEvent): Boolean
-
-    private suspend fun AwaitPointerEventScope.awaitScrollEvent(): PointerEvent {
-        var event: PointerEvent
-        do {
-            event = awaitPointerEvent()
-        } while (event.type != PointerEventType.Scroll)
-        return event
-    }
-
-    private inline val PointerEvent.isConsumed: Boolean get() = changes.fastAny { it.isConsumed }
-    private inline fun PointerEvent.consume() = changes.fastForEach { it.consume() }
-
-}
-
-private class RawMouseWheelScrollNode(
-    scrollingLogicState: State<ScrollingLogic>,
-    mouseWheelScrollConfig: ScrollConfig
-) : MouseWheelScrollNode(scrollingLogicState, mouseWheelScrollConfig) {
-    override fun PointerInputScope.onMouseWheel(pointerEvent: PointerEvent): Boolean {
-        val delta = with(mouseWheelScrollConfig) {
-            calculateMouseWheelScroll(pointerEvent, size)
-        }
-        return scrollingLogic.dispatchRawDelta(delta) != Offset.Zero
-    }
-}
-
-private class AnimatedMouseWheelScrollNode(
-    scrollingLogicState: State<ScrollingLogic>,
-    mouseWheelScrollConfig: ScrollConfig
-) : MouseWheelScrollNode(scrollingLogicState, mouseWheelScrollConfig) {
-    private var isAnimationRunning = false
-    private val channel = Channel<Float>(capacity = Channel.UNLIMITED)
-
-    override fun onAttach() {
-        val density = 1f // TODO: LocalDensity.current.density (no @Composable scope)
-        coroutineScope.launch {
-            while (isActive) {
-                val event = channel.receive()
-                isAnimationRunning = true
-                scrollingLogic.animatedDispatchScroll(event, speed = 1f * density) {
-                    // Sum delta from all pending events to avoid multiple animation restarts.
-                    channel.sumOrNull()
-                }
-                isAnimationRunning = false
-            }
-        }
-    }
-
-    override fun PointerInputScope.onMouseWheel(pointerEvent: PointerEvent): Boolean {
-        val scrollDelta = with(mouseWheelScrollConfig) {
-            calculateMouseWheelScroll(pointerEvent, size)
-        }
-        return if (mouseWheelScrollConfig.isPreciseWheelScroll(pointerEvent)) {
-            // In case of high-resolution wheel, such as a freely rotating wheel with no notches
-            // or trackpads, delta should apply directly without any delays.
-            scrollingLogic.dispatchRawDelta(scrollDelta) != Offset.Zero
-
-            /*
-             * TODO Set isScrollInProgress to true in case of touchpad.
-             *  Dispatching raw delta doesn't cause a progress indication even with wrapping in
-             *  `scrollableState.scroll` block, since it applies the change within single frame.
-             *  Touchpads emit just multiple mouse wheel events, so detecting start and end of this
-             *  "gesture" is not straight forward.
-             *  Ideally it should be resolved by catching real touches from input device instead of
-             *  introducing a timeout (after each event before resetting progress flag).
-             */
-        } else with(scrollingLogic) {
-            val delta = scrollDelta.reverseIfNeeded().toFloat()
-            if (isAnimationRunning) {
-                channel.trySend(delta).isSuccess
-            } else {
-                // Try to apply small delta immediately to conditionally consume
-                // an input event and to avoid useless animation.
-                tryToScrollBySmallDelta(delta, threshold = 4.dp.toPx()) {
-                    channel.trySend(it).isSuccess
-                }
-            }
-        }
-    }
-
-    private fun Channel<Float>.sumOrNull(): Float? {
-        val elements = untilNull { tryReceive().getOrNull() }.toList()
-        return if (elements.isEmpty()) null else elements.sum()
-    }
-
-    private fun <E> untilNull(builderAction: () -> E?) = sequence<E> {
-        do {
-            val element = builderAction()?.also {
-                yield(it)
-            }
-        } while (element != null)
-    }
-
-    private fun ScrollingLogic.tryToScrollBySmallDelta(
-        delta: Float,
-        threshold: Float = 4f,
-        fallback: (Float) -> Boolean
-    ): Boolean {
-        return if (abs(delta) > threshold) {
-            // Gather possibility to scroll by applying a piece of required delta.
-            val testDelta = if (delta > 0f) threshold else -threshold
-            val consumedDelta = scrollableState.dispatchRawDelta(testDelta)
-            consumedDelta != 0f && fallback(delta - testDelta)
-        } else {
-            val consumedDelta = scrollableState.dispatchRawDelta(delta)
-            consumedDelta != 0f
-        }
-    }
-
-    private suspend fun ScrollingLogic.animatedDispatchScroll(
-        eventDelta: Float,
-        speed: Float = 1f,
-        maxDurationMillis: Int = 100,
-        tryReceiveNext: () -> Float?
-    ) {
-        var target = eventDelta
-        tryReceiveNext()?.let {
-            target += it
-        }
-        if (target.isLowScrollingDelta()) {
-            return
-        }
-        scrollableState.scroll {
-            var requiredAnimation = true
-            var lastValue = 0f
-            val anim = AnimationState(0f)
-            while (requiredAnimation) {
-                requiredAnimation = false
-                val durationMillis = (abs(target - anim.value) / speed)
-                    .roundToInt()
-                    .coerceAtMost(maxDurationMillis)
-                anim.animateTo(
-                    target,
-                    animationSpec = tween(
-                        durationMillis = durationMillis,
-                        easing = LinearEasing
-                    ),
-                    sequentialAnimation = true
-                ) {
-                    val delta = value - lastValue
-                    if (!delta.isLowScrollingDelta()) {
-                        val consumedDelta = scrollBy(delta)
-                        if (!(delta - consumedDelta).isLowScrollingDelta()) {
-                            cancelAnimation()
-                            return@animateTo
-                        }
-                        lastValue += delta
-                    }
-                    tryReceiveNext()?.let {
-                        target += it
-                        requiredAnimation = !(target - lastValue).isLowScrollingDelta()
-                        cancelAnimation()
-                    }
-                }
-            }
-        }
-    }
-}
-
 @OptIn(ExperimentalFoundationApi::class)
-private class ScrollingLogic(
+internal class ScrollingLogic(
     val orientation: Orientation,
     val reverseDirection: Boolean,
     val nestedScrollDispatcher: State<NestedScrollDispatcher>,
@@ -678,7 +442,7 @@ private class ScrollDraggableState(
         }
     }
 
-    override suspend fun drag(dragPriority: FoundationMutatePriority, block: suspend DragScope.() -> Unit) {
+    override suspend fun drag(dragPriority: MutatePriority, block: suspend DragScope.() -> Unit) {
         scrollLogic.value.scrollableState.scroll(dragPriority) {
             latestScrollScope = this
             block()
@@ -785,9 +549,3 @@ internal val DefaultScrollMotionDurationScale = object : MotionDurationScale {
     override val scaleFactor: Float
         get() = DefaultScrollMotionDurationScaleFactor
 }
-
-/*
- * Returns true, if the value is too low for visible change in scroll (consumed delta, animation-based change, etc),
- * false otherwise
- */
-private inline fun Float.isLowScrollingDelta(): Boolean = abs(this) < 0.5f
